+++ bt/app/url_tester_window.h
@@ -9,7 +9,7 @@ namespace bt {
         url_tester_window(grey::grey_context& ctx);
 
     private:
-        std::shared_ptr<grey::complex_table<browser_match_result>> tbl;
+        std::shared_ptr<grey::child> w_matches;
         std::vector<std::shared_ptr<browser_match_result>> matches;
         url_payload up;
         url u{""};